chore: rename test files and folder
